# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses/{courseId}/conversations:
    get:
      operationId: listStudyConversations
      tags: [tutoring]
      summary: List tutoring conversations for a course
      parameters:
        - in: path
          name: courseId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Conversation" }
    post:
      operationId: createStudyConversation
      tags: [tutoring]
      summary: Create a tutoring conversation
      parameters:
        - in: path
          name: courseId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConversationInput" }
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
        "404": { description: Course not found }
  /conversations/{conversationId}/messages:
    get:
      operationId: listConversationMessages
      tags: [tutoring]
      summary: Read messages and their source citations
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Conversation messages
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ConversationMessage" }
        "404": { description: Conversation not found }
    post:
      operationId: sendConversationMessage
      tags: [tutoring]
      summary: Ask a question grounded in course sources
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConversationMessageInput" }
      responses:
        "200":
          description: Saved user and assistant messages
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TutorReply" }
        "404": { description: Conversation not found }
        "503": { description: No supporting course material or AI provider unavailable }
...
components:
  schemas:
...
    SourceCitation:
      type: object
      properties:
        unitId: { type: string }
        sourceId: { type: string }
        sourceTitle: { type: string }
        sourceType: { type: string, enum: [pdf, pptx, video] }
        locator:
          type: object
          properties:
            type: { type: string, enum: [page, slide, video] }
            page: { type: integer, minimum: 1 }
            slide: { type: integer, minimum: 1 }
            start_sec: { type: number, minimum: 0 }
            end_sec: { type: number, minimum: 0 }
            imagePath: { type: ["string", "null"] }
          required: [type]
        excerpt: { type: string }
      required: [unitId, sourceId, sourceTitle, sourceType, locator, excerpt]
    ConversationInput:
      type: object
      properties:
        title: { type: string, maxLength: 160 }
    Conversation:
      type: object
      properties:
        id: { type: string }
        courseId: { type: string }
        title: { type: string }
        messageCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, courseId, title, messageCount, createdAt, updatedAt]
    ConversationMessageInput:
      type: object
      properties:
        content: { type: string, minLength: 2, maxLength: 6000 }
      required: [content]
    ConversationMessage:
      type: object
      properties:
        id: { type: string }
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
        citations:
          type: array
          items: { $ref: "#/components/schemas/SourceCitation" }
        createdAt: { type: string, format: date-time }
      required: [id, role, content, citations, createdAt]
    TutorReply:
      type: object
      properties:
        userMessage: { $ref: "#/components/schemas/ConversationMessage" }
        assistantMessage: { $ref: "#/components/schemas/ConversationMessage" }
        supportStatus: { type: string, enum: [supported, insufficient_context] }
      required: [userMessage, assistantMessage, supportStatus]
...